code/requirements.txt
# remember to use
#   pip install -r requirements.txt -f https://download.pytorch.org/whl/torch_stable.html

numpy==1.20.2
pandas==1.1.5
PyYAML==5.4.1

# required for web
flask==1.1.2
flask-restful==0.3.8
flask-cors===3.0.10
gunicorn==20.1.0
pillow==8.2.0

# required for ml-agents
joblib==1.0.1
scikit-learn==0.24.2

# required for NN
torch==1.9.0+cpu
torchvision==0.10.0+cpu
tqdm==4.62

# required for multiprocessing
ray==1.6.0
